refactored extensions
