+++ src/components/index.js
@@ -1,6 +1,6 @@
 import CommunityCard from "./community/CommunityCard";
 import CommunityList from "./community/CommunityList";
-import AutocompleteInputChip from "./input-chip/";
+import AutocompleteInputChip from "./input-chip/AutocompleteInputChip";
 import PlusIconButton from "./common/PlusIconButton";
 import SearchInput from "./common/SearchInput";
 import ThemeSwitchButton from "./themeSwitcher/ThemeSwitchButton";
